+++ src/http/mod.rs
@@ -7,6 +7,7 @@ mod response;
 /// defining the most used status codes
 
 pub mod status_code;
+
 /// exporting all import response implementations
 pub use response::*;